refactor: rename the failover observable scanner to failover scanner, report to failover test report
