+++ docs/adr/index.md
@@ -71,5 +71,5 @@ Full decision text → **[adr.md](adr.md)**
 | [ADR 50 — Metrics Builder Helper — Cheaper Metric Construction on the Recover Path](adr.md#adr-50-metrics-builder-helper-cheaper-metric-construction-on-the-recover-path) | `Metrics` concatenates keys (no `String.format`) + typed `collect` overloads; ~3.6× faster recover-bag build (JMH 744→204 ns/op) | 15-JUN-2026 | Accepted |
 | [ADR 51 — Per-Method Failover Outcome Metric + Method-Identity Threading](adr.md#adr-51-per-method-failover-outcome-metric-method-identity-threading) | `failover.recovery.outcome.total{name,domain,method,outcome}` for failover/recovery/non-recovery rates | 15-JUN-2026 | Accepted — threading mechanism refined by ADR 52 |
 | [ADR 52 — FailoverHandler Method-Aware Contract + AbstractFailoverHandler](adr.md#adr-52-failoverhandler-method-aware-contract-abstractfailoverhandler-refines-adr-51-threading) | Single method-aware `FailoverHandler` contract (`@NonNull Method`); `AbstractFailoverHandler` bridges method-agnostic handlers; method threaded through scatter to slices. Breaking SPI | 15-JUN-2026 | Accepted |
-| [ADR 53 — Overall JaCoCo Coverage Gate](adr.md#adr-53-overall-jacoco-coverage-gate) | Cross-module `jacoco:check` in the `report` module (unpack classes + merge all exec) fails `verify` below 95% line / 95% branch | 15-JUN-2026 | Accepted |
+| [ADR 53 — Overall JaCoCo Coverage Gate](adr.md#adr-53-overall-jacoco-coverage-gate) | Cross-module `jacoco:check` in the `failover-test-report` module (unpack classes + merge all exec) fails `verify` below 95% line / 95% branch | 15-JUN-2026 | Accepted |
 | [ADR 54 — FailoverStore Assembly — Collapse Four Property-Gated Beans into One](adr.md#adr-54-failoverstore-assembly-collapse-four-property-gated-beans-into-one) | Single `failoverStore` bean replaces the 4 `async × multitenant` `` variants; per-tenant-async made explicit; refines ADR 18 | 16-JUN-2026 | Accepted |
