+++ failover-core/src/main/java/com/societegenerale/failover/core/scanner/package-info.java
@@ -3,7 +3,7 @@
  *
  * <p>{@link com.societegenerale.failover.core.scanner.FailoverScanner} is the core
  * interface. The default implementation ({@code SpringContextFailoverScanner} in the
- * {@code failover-observable-scanner} module) locates all methods annotated with
+ * {@code failover-scanner} module) locates all methods annotated with
  * {@link com.societegenerale.failover.annotations.Failover} by walking the Spring
  * {@code ApplicationContext} — no classpath scanning or external libraries required.
  */
