+++ failover-scanner/pom.xml
@@ -3,7 +3,7 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>failover-observable-scanner</artifactId>
+    <artifactId>failover-scanner</artifactId>
     <packaging>jar</packaging>
 
     <parent>
