+++ failover-scanner/src/main/java/com/societegenerale/failover/scanner/package-info.java
@@ -3,4 +3,4 @@
  * Uses the {@link org.springframework.context.ApplicationContext} and Spring reflection utilities
  * to discover {@code @Failover}-annotated methods — no classpath scanning or external libraries required.
  */
-package com.societegenerale.failover.observable.scanner;
+package com.societegenerale.failover.scanner;
